routes/channel: moved channel creation out
